Move DBus objects into their own module
