+++ pepper2/common/drive_types/drive_type.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 from pepper2.daemon.constraint import Constraint
 
 if TYPE_CHECKING:
-    from pepper2.daemon.controller import Controller
+    from pepper2.daemon.dbus.controller import Controller
     from pepper2.common.drives import Drive
