+++ pepper2/daemon/dbus/__init__.py
@@ -1,0 +1,1 @@
+"""D-Bus objects."""
